Изменен тип переменных

+++ PIE-20_Semkina_Lab_2/PIE-20_Semkina_Lab_2/Form1.cs
@@ -19,9 +19,9 @@ namespace PIE_20_Semkina_Lab_2
 
         private void button1_Click(object sender, EventArgs e)
         {
-            int a = Convert.ToInt32(textBox1.Text);
-            int b = Convert.ToInt32(textBox2.Text);
-            int c;
+            double a = Convert.ToInt32(textBox1.Text);
+            double b = Convert.ToInt32(textBox2.Text);
+            double c;
             c = a + b;
             textBox3.Text = Convert.ToString(c);
             label4.Text = "Сложение";
